requirements.txt
pytorch
torchvision
pytorch-lightning
pytorch-lightning-bolts
pycm
